+++ assets/script/incAccess.js
@@ -4,12 +4,14 @@
     doc = paths[1] ?? ""
     book = paths[2] ?? ""
     chapter = paths[3] ?? ""
+    verse = paths[4] ?? "0"
   await fetch("/api/incAccess", {
     method: "post",
     body: JSON.stringify({
       "doc": doc,
       "book": book,
-      "chapter": chapter
+      "chapter": chapter,
+      "verse": verse,
     }),
     headers: {
       "Content-type": "application/json"